finished editProjectAdmin component
