+++ server/src/routes/userRouter.js
@@ -3,6 +3,9 @@ import userApiController from "../controllers/user/userApiController.js";
 import { isLoggedInAPI } from "../middlewares/authMiddleware.js";
 import requireOwnProfile from "../middlewares/requireOwnProfile.js";
 import { requireAdmin, requirePM } from "../middlewares/roleMiddleware.js"; 
+import upload from "../middlewares/multer.js";
+ 
+
 
 const router = Router();
 
@@ -21,6 +24,9 @@ router.delete("/:id", isLoggedInAPI, requireAdmin, userApiController.deleteUser)
 router.put("/workspace", isLoggedInAPI, requirePM, userApiController.editUserWorkspace);
 
 router.put("/:id", isLoggedInAPI, requireOwnProfile, userApiController.editUser);
+
+router.put("/:id/photo", isLoggedInAPI, requireOwnProfile, upload.single("photo"), userApiController.uploadImage);
+
 
 
 /* router.put("/:id", isLoggedInAPI, userApiController.editUser);
